Remove original JS
Because the original JS was minified and likely run through a module
bundler, it is not easy to remove the portions that we do not want.

Instead, we're going to remove everything and add back each piece we
need to keep by manually transforming it back into human-readable source
code.